Refactor lifecyle.tsx
