+++ ipaas/src/components/Lifecycle/LCStateDiagram.tsx
@@ -1,0 +1,372 @@
+/**
+ * Copyright (c) 2026, WSO2 LLC. (https://www.wso2.com).
+ *
+ * WSO2 LLC. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+import type { JSX } from 'react';
+import type { LifecycleStateTransition } from '../../api/apim';
+import { EVENT_TARGET } from '../../constants/lifecycle';
+import type { ApiState, BubbleStyle } from '../../types/lifecycle';
+
+function getMappedTargetForEvent(event: string): ApiState {
+  return EVENT_TARGET[event] ?? 'Created';
+}
+
+function getBubbleStyle(bubbleState: ApiState, currentState: string, availableTransitions: LifecycleStateTransition[]): BubbleStyle {
+  if (currentState === bubbleState) return 'SELECTED';
+  const enabled = availableTransitions.some((t) => getMappedTargetForEvent(t.event) === bubbleState);
+  return enabled ? 'ENABLED' : 'DISABLED';
+}
+
+const HEX_PATH =
+  'M56,3.46410162 L87.3012702,21.5358984 C91.0140831,23.6794919 93.3012702,27.6410162 93.3012702,31.9282032 L93.3012702,68.0717968 C93.3012702,72.3589838 91.0140831,76.3205081 87.3012702,78.4641016 L56,96.5358984 C52.2871871,98.6794919 47.7128129,98.6794919 44,96.5358984 L12.6987298,78.4641016 C8.98591689,76.3205081 6.69872981,72.3589838 6.69872981,68.0717968 L6.69872981,31.9282032 C6.69872981,27.6410162 8.98591689,23.6794919 12.6987298,21.5358984 L44,3.46410162 C47.7128129,1.32050808 52.2871871,1.32050808 56,3.46410162 Z';
+
+const SHADOW_FILTER = (
+  <filter x="-53.3%" y="-58.1%" width="206.6%" height="216.3%" filterUnits="objectBoundingBox" id="filter-2">
+    <feOffset dx="0" dy="0" in="SourceAlpha" result="shadowOffsetOuter1" />
+    <feGaussianBlur stdDeviation="20" in="shadowOffsetOuter1" result="shadowBlurOuter1" />
+    <feColorMatrix values="0 0 0 0 0   0 0 0 0 0   0 0 0 0 0  0 0 0 0.2 0" type="matrix" in="shadowBlurOuter1" />
+  </filter>
+);
+
+function PrototypedBubble({ style }: { style: BubbleStyle }) {
+  const iconPath =
+    'M13,0.25 C14.5187831,0.25 15.75,1.48121694 15.75,3 L15.75,3 L15.75,13 C15.75,14.5187831 14.5187831,15.75 13,15.75 L13,15.75 L3,15.75 C1.48121694,15.75 0.25,14.5187831 0.25,13 L0.25,13 L0.25,3 C0.25,1.48121694 1.48121694,0.25 3,0.25 L3,0.25 Z M7.25,1.75 L3,1.75 C2.30964406,1.75 1.75,2.30964406 1.75,3 L1.75,3 L1.75,13 C1.75,13.6903559 2.30964406,14.25 3,14.25 L3,14.25 L7.25,14.25 L7.25,1.75 Z M14.25,8.75 L8.75,8.75 L8.75,14.25 L13,14.25 C13.6472087,14.25 14.1795339,13.7581253 14.2435464,13.1278052 L14.25,13 L14.25,8.75 Z M13,1.75 L8.75,1.75 L8.75,7.25 L14.25,7.25 L14.25,3 C14.25,2.35279131 13.7581253,1.8204661 13.1278052,1.75645361 L13,1.75 Z';
+  const isSelected = style === 'SELECTED';
+  const isDisabled = style === 'DISABLED';
+  const strokeColor = isDisabled ? '#CBCEDB' : '#2F5FA7';
+  const iconColor = isDisabled ? '#CBCEDB' : isSelected ? '#FFFFFF' : '#2F5FA7';
+  const textColor = isSelected ? '#FFFFFF' : '#40404B';
+  return (
+    <g id="Group-7" transform="translate(129.499894, 0.000000)" data-testid={`${style.toLowerCase()}-prototyped-bubble`}>
+      {isSelected ? (
+        <g id="Polygon-Copy-3">
+          <use fill="black" fillOpacity="1" filter="url(#filter-2)" xlinkHref="#path-1" />
+          <use fill="#2F5FA7" fillRule="evenodd" xlinkHref="#path-1" />
+        </g>
+      ) : (
+        <path d={HEX_PATH} id="Polygon-Copy-3" stroke={strokeColor} fill="#FFFFFF" />
+      )}
+      <text id="Prototyped" fontSize="11" fontWeight="500" fill={textColor}>
+        <tspan x="17.0" y="65">
+          Pre-released
+        </tspan>
+      </text>
+      <g id="Icon-Prototyped" transform="translate(42.000000, 31.000000)" fill={iconColor} fillRule="nonzero">
+        <path d={iconPath} id="Combined-Shape" />
+      </g>
+    </g>
+  );
+}
+
+function CreatedBubble({ style }: { style: BubbleStyle }) {
+  const iconPath =
+    'M13,0.25 C14.5187831,0.25 15.75,1.48121694 15.75,3 L15.75,9.17157288 C15.75,9.90091822 15.4602687,10.6003915 14.9445436,11.1161165 L12.1161165,14.9445436 C11.6003915,15.4602687 10.9009182,15.75 10.1715729,15.75 L3,15.75 C1.48121694,15.75 0.25,14.5187831 0.25,13 L0.25,3 C0.25,1.48121694 1.48121694,0.25 3,0.25 L13,0.25 Z M13,1.75 L3,1.75 C2.30964406,1.75 1.75,2.30964406 1.75,3 L1.75,13 C1.75,13.6903559 2.30964406,14.25 3,14.25 L10.1715729,14.25 C10.5030935,14.25 10.8210359,14.1183039 11.0554564,13.8838835 L13.8838835,10.0554564 C14.1183039,9.82103588 14.25,9.50309349 14.25,9.17157288 L14.25,3 C14.25,2.30964406 13.6903559,1.75 13,1.75 Z M10.4921525,4.95634663 C10.7568202,4.70411797 11.1733062,4.67702559 11.4702266,4.90861577 C11.7968391,5.16336497 11.8550958,5.63465174 11.6003466,5.96126423 L11.6003466,5.96126423 L7.56857841,11.1303712 L4.46954215,8.03023607 L4.39693891,7.94610477 C4.1791365,7.65245466 4.20341641,7.23579528 4.46973016,6.96957592 L4.46973016,6.96957592 L4.55386147,6.89697267 C4.84751157,6.67917026 5.26417096,6.70345017 5.53039032,6.96976393 L5.53039032,6.96976393 L7.42996238,8.869 L10.4175782,5.03873577 Z';
+  const isSelected = style === 'SELECTED';
+  const isDisabled = style === 'DISABLED';
+  const strokeColor = isDisabled ? '#CBCEDB' : '#2F5FA7';
+  const iconColor = isDisabled ? '#CBCEDB' : isSelected ? '#FFFFFF' : '#2F5FA7';
+  const textColor = isSelected ? '#FFFFFF' : '#40404B';
+  return (
+    <g id="Group-6" transform="translate(129.499894, 312.000000)" data-testid={`${style.toLowerCase()}-created-bubble`}>
+      {isSelected ? (
+        <g id="Polygon-Copy-4">
+          <use fill="black" fillOpacity="1" filter="url(#filter-2)" xlinkHref="#path-1" />
+          <use fill="#2F5FA7" fillRule="evenodd" xlinkHref="#path-1" />
+        </g>
+      ) : (
+        <path d={HEX_PATH} id="Polygon-Copy-4" stroke={strokeColor} fill="#FFFFFF" />
+      )}
+      <text id="Created" fontSize="11" fontWeight="500" fill={textColor}>
+        <tspan x="28.4516602" y="65">
+          Created
+        </tspan>
+      </text>
+      <g id="Icon-Created" transform="translate(42.000000, 31.000000)" fill={iconColor} fillRule="nonzero">
+        <path d={iconPath} id="Combined-Shape" />
+      </g>
+    </g>
+  );
+}
+
+function PublishedBubble({ style }: { style: BubbleStyle }) {
+  const iconPath =
+    'M8.00364273,0.25 C7.12225346,0.25 6.19350787,0.534415244 5.52653937,1.04255832 C4.35089132,1.94832995 3.60248426,3.17904731 3.60248426,5.13752012 L3.60861623,5.42187856 C3.67779503,7.00811337 4.32691514,8.17098708 5.28130824,9.10651368 L5.328,9.151 L5.30453945,9.2051176 C5.21066869,9.38806103 5.00569343,9.57531574 4.60956113,9.75046822 L3.36085868,10.2674783 L2.82417286,10.4973925 C1.56311345,11.0574961 0.896291331,11.5227551 0.583488553,12.2367286 C-0.153129676,13.9180623 0.716820608,15.75 3.00527715,15.75 L13.0052772,15.75 C15.3816926,15.75 16.3677199,13.6905409 15.3566392,12.1372936 C14.9378748,11.4939772 14.1684255,11.0031728 12.8584606,10.4026455 L11.3872516,9.75846698 C11.1252787,9.64341662 10.9458102,9.51258167 10.8287769,9.37341823 L10.7567907,9.27579964 C10.7366468,9.24399435 10.7201879,9.21288667 10.7069147,9.18236665 L10.694,9.147 L10.7510232,9.09323648 C11.7412714,8.1050459 12.3872896,6.8761502 12.3872896,5.1462755 C12.3872896,3.18244531 11.6269637,1.90026537 10.4865386,1.04031227 C9.80080002,0.536671248 8.87124742,0.25 8.00364273,0.25 Z M8.00364273,1.75 C8.56057621,1.75 9.17820231,1.94047405 9.59106384,2.24365527 C10.3747492,2.83467961 10.8872896,3.69900496 10.8872896,5.1462755 C10.8872896,6.55426326 10.3062132,7.50402489 9.41619844,8.28929591 L9.16240183,8.51322379 L9.16171689,8.84850965 C9.13843029,9.33257802 9.26140111,9.84019273 9.68077235,10.338864 C9.91170671,10.6134664 10.2144027,10.8493141 10.5901391,11.0401953 L10.8234622,11.1490098 L11.9785821,11.6510871 L12.4299571,11.8583063 C13.3510911,12.2951667 13.9033815,12.6543029 14.0995177,12.9556122 C14.4835967,13.5456441 14.1463654,14.25 13.0052772,14.25 L3.00527715,14.25 C1.91849056,14.25 1.61712304,13.6153815 1.95741388,12.8386664 C2.05494223,12.6160576 2.43825074,12.3431986 3.13461264,12.0084731 L3.50900763,11.8359011 L5.21442969,11.1231046 C6.38721888,10.6045507 6.85951293,9.76999941 6.87188414,8.93790882 L6.86694453,8.51398695 L6.61192059,8.29330331 C5.70592595,7.50930558 5.10248426,6.54120366 5.10248426,5.13752012 C5.10248426,3.70157881 5.60448905,2.87605746 6.43879893,2.23325813 C6.82848128,1.93638476 7.4371126,1.75 8.00364273,1.75 Z';
+  const isSelected = style === 'SELECTED';
+  const isDisabled = style === 'DISABLED';
+  const strokeColor = isDisabled ? '#CBCEDB' : '#2F5FA7';
+  const iconColor = isDisabled ? '#CBCEDB' : isSelected ? '#FFFFFF' : '#2F5FA7';
+  const textColor = isSelected ? '#FFFFFF' : isDisabled ? '#40404B' : '#2F5FA7';
+  return (
+    <g id="Group-3" transform="translate(303.499894, 153.000000)" data-testid={`${style.toLowerCase()}-published-bubble`}>
+      {isSelected ? (
+        <g id="Polygon-Copy-5">
+          <use fill="black" fillOpacity="1" filter="url(#filter-2)" xlinkHref="#path-1" />
+          <use fill="#2F5FA7" fillRule="evenodd" xlinkHref="#path-1" />
+        </g>
+      ) : (
+        <path d={HEX_PATH} id="Polygon-Copy-5" stroke={strokeColor} fill="#FFFFFF" />
+      )}
+      <text id="Published" fontSize="11" fontWeight="500" fill={textColor}>
+        <tspan x="24.1247559" y="65">
+          Published
+        </tspan>
+      </text>
+      <g id="Icon-Published" transform="translate(42.000000, 31.000000)" fill={iconColor} fillRule="nonzero">
+        <path d={iconPath} id="person" />
+      </g>
+    </g>
+  );
+}
+
+function DeprecatedBubble({ style }: { style: BubbleStyle }) {
+  const iconPath =
+    'M8,0.2409091 C8.9181734,0.2409091 9.67119234,0.948019373 9.7441988,1.84738189 L9.75,1.9909091 L9.75,2.5049091 L13.25,2.50531085 C13.6642136,2.50531085 14,2.84109729 14,3.25531085 C14,3.66952442 13.6642136,4.00531085 13.25,4.00531085 L12.75,4.0049091 L12.75,14 C12.75,14.9181734 12.0428897,15.6711923 11.1435272,15.7441988 L11,15.75 L3,15.75 C2.0818266,15.75 1.32880766,15.0428897 1.2558012,14.1435272 L1.25,14 L1.25,4.0049091 L0.75,4.00531085 C0.335786438,4.00531085 5.07265313e-17,3.66952442 0,3.25531085 C-5.07265313e-17,2.84109729 0.335786438,2.50531085 0.75,2.50531085 L4.25,2.5049091 L4.25,1.9909091 C4.25,1.0727357 4.95711027,0.31971676 5.85647279,0.246710299 L6,0.2409091 L8,0.2409091 Z M11.25,4.0049091 L2.75,4.0049091 L2.75,14 C2.75,14.1183467 2.83223341,14.2174868 2.94267729,14.2433973 L3,14.25 L11,14.25 C11.1183467,14.25 11.2174868,14.1677666 11.2433973,14.0573227 L11.25,14 L11.25,4.0049091 Z M7,5.4909091 C7.37969577,5.4909091 7.69349096,5.77306298 7.74315338,6.13913854 L7.75,6.2409091 L7.75,12 C7.75,12.4142136 7.41421356,12.75 7,12.75 C6.62030423,12.75 6.30650904,12.4678461 6.25684662,12.1017706 L6.25,12 L6.25,6.2409091 C6.25,5.82669554 6.58578644,5.4909091 7,5.4909091 Z M8,1.7409091 L6,1.7409091 C5.88165327,1.7409091 5.78251318,1.82314251 5.75660268,1.93358639 L5.75,1.9909091 L5.75,2.5049091 L8.25,2.5049091 L8.25,1.9909091 C8.25,1.87256237 8.16776659,1.77342228 8.05732271,1.74751178 L8,1.7409091 Z';
+  const isSelected = style === 'SELECTED';
+  const isDisabled = style === 'DISABLED';
+  const strokeColor = isDisabled ? '#CBCEDB' : '#2F5FA7';
+  const iconColor = isDisabled ? '#CBCEDB' : isSelected ? '#FFFFFF' : '#2F5FA7';
+  const textColor = isSelected ? '#FFFFFF' : '#40404B';
+  return (
+    <g id="Group-5" transform="translate(481.499894, 153.000000)" data-testid={`${style.toLowerCase()}-deprecated-bubble`}>
+      {isSelected ? (
+        <g id="Polygon-Copy-6">
+          <use fill="black" fillOpacity="1" filter="url(#filter-2)" xlinkHref="#path-1" />
+          <use fill="#2F5FA7" fillRule="evenodd" xlinkHref="#path-1" />
+        </g>
+      ) : (
+        <path d={HEX_PATH} id="Polygon-Copy-6" stroke={strokeColor} fill="#FFFFFF" />
+      )}
+      <text id="Deprecated" fontSize="11" fontWeight="500" fill={textColor}>
+        <tspan x="17.7890625" y="65">
+          Deprecated
+        </tspan>
+      </text>
+      <g id="Icon-Deprecated" transform="translate(43.000000, 31.000000)" fill={iconColor} fillRule="nonzero">
+        <path d={iconPath} id="Combined-Shape" />
+      </g>
+    </g>
+  );
+}
+
+function RetiredBubble({ style }: { style: BubbleStyle }) {
+  const iconPath =
+    'M12.25,0 C13.1681734,0 13.9211923,0.707110273 13.9941988,1.60647279 L14,1.75 L14,11.75 C14,12.6681734 13.2928897,13.4211923 12.3935272,13.4941988 L12.25,13.5 L10,13.5 L10,14.3069995 C10,15.2251729 9.29288973,15.9781919 8.39352721,16.0511983 L8.25,16.0569995 C8.04015409,16.0569995 7.83201883,16.0192574 7.63553398,15.9455756 L7.63553398,15.9455756 L1.63553398,13.6955756 C0.952502459,13.4394388 0.5,12.7864775 0.5,12.0569995 L0.5,12.0569995 L0.5,1.75 C0.5,0.831826603 1.20711027,0.0788076603 2.10647279,0.00580119953 L2.25,0 L12.25,0 Z M2.25,1.94300047 C2.11192881,1.94300047 2,2.05492928 2,2.19300047 L2,2.19300047 L2,12.0569995 C2,12.1612107 2.06464321,12.2544909 2.16221914,12.2910818 L2.16221914,12.2910818 L8.16221914,14.5410818 C8.1902884,14.5516078 8.22002201,14.5569995 8.25,14.5569995 C8.38807119,14.5569995 8.5,14.4450707 8.5,14.3069995 L8.5,14.3069995 L8.5,4.44300047 C8.5,4.33878933 8.43535679,4.24550915 8.33778086,4.20891817 L8.33778086,4.20891817 L2.33778086,1.95891817 C2.3097116,1.9483922 2.27997799,1.94300047 2.25,1.94300047 Z M12.25,1.5 L5.387,1.5 L8.86446602,2.80442441 C9.50196211,3.04348544 9.93864177,3.62822465 9.9940324,4.29822 L10,4.44300047 L10,12 L12.25,12 C12.3683467,12 12.4674868,11.9177666 12.4933973,11.8073227 L12.5,11.75 L12.5,1.75 C12.5,1.63165327 12.4177666,1.53251318 12.3073227,1.50660268 L12.25,1.5 Z M6.75,7 C7.12969577,7 7.44349096,7.28215388 7.49315338,7.64822944 L7.5,7.75 L7.5,9.75 C7.5,10.1642136 7.16421356,10.5 6.75,10.5 C6.37030423,10.5 6.05650904,10.2178461 6.00684662,9.85177056 L6,9.75 L6,7.75 C6,7.33578644 6.33578644,7 6.75,7 Z';
+  const isSelected = style === 'SELECTED';
+  const isDisabled = style === 'DISABLED';
+  const strokeColor = isDisabled ? '#CBCEDB' : '#2F5FA7';
+  const iconColor = isDisabled ? '#CBCEDB' : isSelected ? '#FFFFFF' : '#2F5FA7';
+  const textColor = isSelected ? '#FFFFFF' : '#40404B';
+  return (
+    <g id="Group-4" transform="translate(659.499894, 153.000000)" data-testid={`${style.toLowerCase()}-retired-bubble`}>
+      {isSelected ? (
+        <g id="Polygon-Copy-7">
+          <use fill="black" fillOpacity="1" filter="url(#filter-2)" xlinkHref="#path-1" />
+          <use fill="#2F5FA7" fillRule="evenodd" xlinkHref="#path-1" />
+        </g>
+      ) : (
+        <path d={HEX_PATH} id="Polygon-Copy-7" stroke={strokeColor} fill="#FFFFFF" />
+      )}
+      <text id="Retired" fontSize="11" fontWeight="500" fill={textColor}>
+        <tspan x="30.2714844" y="65">
+          Retired
+        </tspan>
+      </text>
+      <g id="Icon-Retired" transform="translate(43.000000, 31.000000)" fill={iconColor} fillRule="nonzero">
+        <path d={iconPath} id="Combined-Shape" />
+      </g>
+    </g>
+  );
+}
+
+function ArrowsFromPrototyped({ state }: { state: string }) {
+  if (state === 'Prototyped') {
+    return (
+      <g>
+        <path
+          fillRule="nonzero"
+          fill="#2F5FA7"
+          id="Path"
+          d="m262.96215,14.46958c32.99347,17.10919 59.99347,44.15291 77.04938,77.15808l0.26496,0.5163l3.61727,-1.73956l-0.15349,10.06114l-7.95681,-6.15947l3.59103,-1.72811l-0.25136,-0.49123c-16.78967,-32.48995 -43.27003,-59.17151 -75.62892,-76.21082l-0.99241,-0.51859l0.46035,-0.88774z"
+        />
+        <path
+          fillRule="nonzero"
+          fill="#2F5FA7"
+          id="Path"
+          d="m115.76661,51.70504l0.46419,0.88574c-27.52079,14.42275 -49.43838,37.91668 -61.87832,66.40015c-7.45254,17.06394 -11.34973,35.54869 -11.34973,54.58141c0,25.69318 7.10965,50.33816 20.33402,71.71402c11.43347,18.48105 27.14844,34.02146 45.74898,45.24442l0.417,0.24956l1.93176,-3.4829l5.68909,8.29965l-10.05317,-0.42851l1.94732,-3.51324l-0.4486,-0.26833c-18.73643,-11.30495 -34.56582,-26.95852 -46.0828,-45.57453c-13.32146,-21.53281 -20.4836,-46.35975 -20.4836,-72.24014c0,-19.17112 3.9259,-37.79207 11.43332,-54.98164c12.53114,-28.69231 34.6084,-52.35738 62.33054,-66.88566z"
+        />
+      </g>
+    );
+  }
+  return (
+    <g>
+      <path
+        d="M112.717591,286.954981 L117.546407,294.76091 L118.014038,295.516849 L117.125178,295.523936 L107.94669,295.597116 L107.938717,294.597148 L116.227894,294.53 L111.867159,287.481065 L112.717591,286.954981 Z M113.439552,292.420368 L114.315009,292.903665 L113.831712,293.779122 L112.956255,293.295825 L113.439552,292.420368 Z M109.100396,289.969004 L109.956692,290.48549 L109.440206,291.341786 L108.583911,290.8253 L109.100396,289.969004 Z M104.876971,287.315194 L105.713836,287.862602 L105.166427,288.699468 L104.329562,288.152059 L104.876971,287.315194 Z M100.752523,284.508585 L101.568871,285.086146 L100.99131,285.902494 L100.174963,285.324933 L100.752523,284.508585 Z M96.734561,281.554676 L97.5293437,282.16157 L96.9224496,282.956353 L96.1276669,282.349459 L96.734561,281.554676 Z M92.8552903,278.48005 L93.6276521,279.115233 L92.9924693,279.887595 L92.2201075,279.252412 L92.8552903,278.48005 Z M89.0554909,275.237422 L89.8042271,275.90029 L89.1413589,276.649027 L88.3926228,275.986159 L89.0554909,275.237422 Z M85.3776539,271.860124 L86.101797,272.549774 L85.4121471,273.273917 L84.6880041,272.584267 L85.3776539,271.860124 Z M81.8255822,268.351957 L82.5241849,269.067467 L81.8086751,269.766069 L81.1100724,269.05056 L81.8255822,268.351957 Z M78.4085226,264.722721 L79.0807006,265.463111 L78.340311,266.135289 L77.668133,265.394899 L78.4085226,264.722721 Z M75.1322331,260.979905 L75.7771472,261.74416 L75.0128922,262.389074 L74.367978,261.624819 L75.1322331,260.979905 Z M71.9819803,257.105807 L72.5986527,257.893027 L71.8114329,258.509699 L71.1947604,257.72248 L71.9819803,257.105807 Z M68.9796,253.126066 L69.5672301,253.935196 L68.7581005,254.522826 L68.1704703,253.713696 L68.9796,253.126066 Z M66.1255489,249.041903 L66.6833362,249.871887 L65.8533523,250.429674 L65.2955649,249.59969 L66.1255489,249.041903 Z M63.4228604,244.857598 L63.9500277,245.707359 L63.1002662,246.234527 L62.5730989,245.384765 L63.4228604,244.857598 Z M60.8681965,240.566294 L61.3639103,241.43478 L60.4954244,241.930493 L59.9997105,241.062008 L60.8681965,240.566294 Z M58.4744216,236.188093 L58.9379918,237.074153 L58.0519316,237.537724 L57.5883614,236.651663 L58.4744216,236.188093 Z M56.241696,231.722872 L56.6724305,232.625351 L55.7699518,233.056085 L55.3392173,232.153606 L56.241696,231.722872 Z M54.1764395,227.183128 L54.5737405,228.100816 L53.6560521,228.498117 L53.2587511,227.580429 L54.1764395,227.183128 Z M52.278891,222.569233 L52.6421694,223.500914 L51.7104888,223.864192 L51.3472104,222.932511 L52.278891,222.569233 Z M50.5603162,217.911324 L50.8892158,218.855689 L49.9448509,219.184589 L49.6159513,218.240224 L50.5603162,217.911324 Z M49.0107415,213.184266 L49.3047349,214.140073 L48.3489275,214.434067 L48.054934,213.478259 L49.0107415,213.184266 Z M47.6303695,208.38309 L47.8889116,209.34909 L46.9229116,209.607632 L46.6643695,208.641632 L47.6303695,208.38309 Z M46.4288781,203.53832 L46.6516728,204.513185 L45.6768074,204.73598 L45.4540127,203.761114 L46.4288781,203.53832 Z M45.4077145,198.659086 L45.594561,199.641475 L44.6121719,199.828322 L44.4253254,198.845933 L45.4077145,198.659086 Z M44.56828,193.759242 L44.7191111,194.747802 L43.7305515,194.898633 L43.5797205,193.910073 L44.56828,193.759242 Z M43.9056915,188.810868 L44.0202707,189.804282 L43.0268566,189.918861 L42.9122774,188.925447 L43.9056915,188.810868 Z M43.4244595,183.845868 L43.5028236,184.842792 L42.5058988,184.921157 L42.4275347,183.924232 L43.4244595,183.845868 Z M43.1237685,178.868607 L43.1660273,179.867714 L42.1669206,179.909972 L42.1246618,178.910866 L43.1237685,178.868607 Z M43.0027224,173.888107 L43.0090932,174.888087 L42.0091135,174.894457 L42.0027427,173.894478 L43.0027224,173.888107 Z M42.060688,168.871689 L43.0602535,168.901166 L43.030777,169.900731 L42.0312115,169.871254 L42.060688,168.871689 Z M42.2991605,163.855549 L43.2970117,163.921071 L43.2314899,164.918922 L42.2336388,164.853401 L42.2991605,163.855549 Z M42.7156406,158.884324 L43.7104809,158.985778 L43.6090269,159.980618 L42.6141866,159.879164 L42.7156406,158.884324 Z M43.3174174,153.888876 L44.3078869,154.026608 L44.1701548,155.017077 L43.1796853,154.879345 L43.3174174,153.888876 Z M44.1009549,148.92272 L45.0857129,149.09665 L44.9117831,150.081408 L43.9270251,149.907478 L44.1009549,148.92272 Z M45.0658496,143.990192 L46.0435556,144.20017 L45.833577,145.177876 L44.855871,144.967898 L45.0658496,143.990192 Z M46.2065543,139.115648 L47.1759087,139.361315 L46.9302421,140.330669 L45.9608877,140.085002 L46.2065543,139.115648 Z M47.5318607,134.26355 L48.4915026,134.544775 L48.2102776,135.504417 L47.2506357,135.223191 L47.5318607,134.26355 Z M49.0345913,129.4652 L49.9832173,129.7816 L49.6668177,130.730226 L48.7181917,130.413826 L49.0345913,129.4652 Z M50.7097062,124.734575 L51.6460564,125.085643 L51.2949892,126.021993 L50.358639,125.670926 L50.7097062,124.734575 Z M52.5585397,120.066412 L53.4813517,120.451663 L53.0961012,121.374475 L52.1732892,120.989224 L52.5585397,120.066412 Z M54.5629057,115.501269 L55.4710573,115.91991 L55.0524159,116.828062 L54.1442642,116.40942 L54.5629057,115.501269 Z M56.7524863,110.972263 L57.6446453,111.423985 L57.1929237,112.316144 L56.3007647,111.864422 L56.7524863,110.972263 Z M59.1092255,106.522381 L59.984198,107.006554 L59.5000255,107.881527 L58.625053,107.397354 L59.1092255,106.522381 Z M61.6251144,102.166115 L62.48177,102.682003 L61.9658813,103.538659 L61.1092257,103.02277 L61.6251144,102.166115 Z M64.2969232,97.9075579 L65.1341583,98.454401 L64.5873152,99.2916361 L63.7500801,98.744793 L64.2969232,97.9075579 Z M67.1246768,93.74619 L67.9413892,94.323235 L67.3643443,95.1399474 L66.5476319,94.5629025 L67.1246768,93.74619 Z M70.0987035,89.6955637 L70.8938609,90.3019668 L70.2874579,91.0971242 L69.4923004,90.4907212 L70.0987035,89.6955637 Z M73.215212,85.7594671 L73.9878073,86.3943659 L73.3529085,87.1669612 L72.5803132,86.5320624 L73.215212,85.7594671 Z M76.4779223,81.9331971 L77.2269175,82.5957726 L76.564342,83.3447677 L75.8153468,82.6821922 L76.4779223,81.9331971 Z M79.8448929,78.2624372 L80.5695472,78.9515498 L79.8804346,79.6762041 L79.1557803,78.9870915 L79.8448929,78.2624372 Z M83.3776762,74.680194 L84.0767968,75.3951978 L83.361793,76.0943184 L82.6626724,75.3793146 L83.3776762,74.680194 Z M87.038653,71.2285082 L87.7113106,71.9684621 L86.9713567,72.6411197 L86.2986991,71.9011658 L87.038653,71.2285082 Z M90.8196655,67.9148658 L91.4649804,68.6787824 L90.7010638,69.3240974 L90.0557489,68.5601807 L90.8196655,67.9148658 Z M94.716487,64.7426218 L95.3336006,65.5294959 L94.5467266,66.1466095 L93.929613,65.3597354 L94.716487,64.7426218 Z M98.7111377,61.7251272 L99.299312,62.5338614 L98.4905778,63.1220357 L97.9024035,62.3133015 L98.7111377,61.7251272 Z M102.831346,58.8419764 L103.389633,59.6716242 L102.559985,60.2299114 L102.001698,59.4002636 L102.831346,58.8419764 Z M107.050255,56.1134756 L107.577894,56.9629443 L106.728425,57.4905833 L106.200786,56.6411146 L107.050255,56.1134756 Z M111.363381,53.5423737 L111.859637,54.4105497 L110.991461,54.9068062 L110.495205,54.0386302 L111.363381,53.5423737 Z M115.763758,51.1326991 L116.227945,52.0184366 L115.342207,52.482623 L114.878021,51.5968854 L115.763758,51.1326991 Z"
+        id="Path"
+        fill="#8D91A3"
+      />
+      <path
+        d="M342.895899,90.3540473 L343.893762,90.4193916 L343.29398,99.5785544 L343.235897,100.465544 L342.508038,99.9553024 L334.992082,94.6864881 L335.566104,93.8676477 L341.902894,98.31 L342.389894,98.076 L342.895899,90.3540473 Z M340.436227,93.4683861 L340.868488,94.3701346 L339.96674,94.8023956 L339.534479,93.900647 L340.436227,93.4683861 Z M338.172522,88.9935686 L338.636239,89.8795522 L337.750255,90.3432687 L337.286539,89.4572851 L338.172522,88.9935686 Z M335.75807,84.592658 L336.246435,85.4652972 L335.373796,85.9536627 L334.885431,85.0810236 L335.75807,84.592658 Z M333.225402,80.2678192 L333.737965,81.1264688 L332.879316,81.6390317 L332.366753,80.7803821 L333.225402,80.2678192 Z M330.569275,76.0096593 L331.105638,76.8536471 L330.26165,77.3900094 L329.725288,76.5460216 L330.569275,76.0096593 Z M327.794819,71.8266817 L328.354537,72.6553645 L327.525855,73.2150831 L326.966136,72.3864003 L327.794819,71.8266817 Z M324.911831,67.7325159 L325.49439,68.5453048 L324.681601,69.1278632 L324.099043,68.3150743 L324.911831,67.7325159 Z M321.910451,63.7127021 L322.515417,64.5089535 L321.719166,65.1139194 L321.1142,64.3176679 L321.910451,63.7127021 Z M318.796325,59.775463 L319.423224,60.5545635 L318.644123,61.1814625 L318.017224,60.402362 L318.796325,59.775463 Z M315.571242,55.9233746 L316.21959,56.6847189 L315.458245,57.3330667 L314.809898,56.5717224 L315.571242,55.9233746 Z M312.241617,52.1640893 L312.91089,52.9071053 L312.167875,53.576379 L311.498601,52.833363 L312.241617,52.1640893 Z M308.804944,48.4949017 L309.49464,49.2190004 L308.770542,49.9086968 L308.080845,49.1845981 L308.804944,48.4949017 Z M305.297211,44.9519835 L306.00663,45.6567709 L305.301842,46.3661895 L304.592424,45.661402 L305.297211,44.9519835 Z M301.672931,41.4885433 L302.401636,42.1733713 L301.716808,42.902076 L300.988103,42.217248 L301.672931,41.4885433 Z M297.948171,38.1219845 L298.695633,38.7862893 L298.031328,39.5337511 L297.283866,38.8694463 L297.948171,38.1219845 Z M294.13005,34.8593232 L294.895704,35.502576 L294.252451,36.2682299 L293.486797,35.6249771 L294.13005,34.8593232 Z M290.226194,31.70701 L291.009442,32.3287197 L290.387732,33.1119675 L289.604484,32.4902578 L290.226194,31.70701 Z M286.233696,28.662743 L287.033954,29.2623981 L286.434299,30.0626566 L285.634041,29.4630015 L286.233696,28.662743 Z M282.154788,25.7287366 L282.971463,26.3058339 L282.394366,27.1225093 L281.57769,26.545412 L282.154788,25.7287366 Z M277.997439,22.9109771 L278.829904,23.4650535 L278.275828,24.2975193 L277.443362,23.7434429 L277.997439,22.9109771 Z M273.758367,20.2075915 L274.606007,20.7381625 L274.075436,21.585803 L273.227796,21.055232 L273.758367,20.2075915 Z M269.44575,17.6242383 L270.307918,18.1308617 L269.801294,18.9930291 L268.939127,18.4864057 L269.44575,17.6242383 Z M265.128742,15.1988367 L266.004574,15.6814535 L265.521957,16.5572851 L264.646125,16.0746684 L265.128742,15.1988367 Z"
+        id="Path"
+        fill="#8D91A3"
+      />
+    </g>
+  );
+}
+
+function ArrowsFromPublished({ state }: { state: string }) {
+  if (state === 'Published') {
+    return (
+      <g>
+        <polygon id="Path-3" fill="#2F5FA7" points="453.499894 168 462.499894 172.5 453.499894 177 453.499894 173 420.499894 173 420.499894 172 453.499894 172" />
+        <path
+          d="M333.743853,262.89688 L334.607349,263.401235 C318.941153,290.222994 296.520448,312.609515 269.684781,328.233618 L269.231894,328.494 L271.153024,332.018855 L261.097055,332.375925 L266.84498,324.116909 L268.752894,327.616 L269.191532,327.363654 C295.873671,311.826642 318.166641,289.566291 333.743853,262.89688 Z"
+          id="Path"
+          fill="#2F5FA7"
+        />
+        <path
+          d="M242.513842,51.302502 L252.564636,51.7836846 L250.597894,55.286 L250.939521,55.4921331 C274.912469,70.0976174 293.957657,91.7620737 305.315246,117.446466 L305.675108,118.267542 L304.757902,118.665955 C293.527463,92.8118211 274.471499,70.9999372 250.419243,56.3461297 L250.107894,56.158 L248.159373,59.6318469 L242.513842,51.302502 Z"
+          id="Path"
+          fill="#2F5FA7"
+        />
+      </g>
+    );
+  }
+  return (
+    <g>
+      <path
+        fillRule="nonzero"
+        fill="#8D91A3"
+        id="Path-3"
+        d="m453.91183,167.56489l8,4.5l0.77473,0.43579l-0.77473,0.43579l-8,4.5l-0.49026,-0.87158l6.33613,-3.56421l-0.091,0l0,-1l0.091,0l-6.33613,-3.56421l0.49026,-0.87158zm-33.24513,4.43579l0,1l-1,0l0,-1l1,0zm5,0l0,1l-1,0l0,-1l1,0zm5,0l0,1l-1,0l0,-1l1,0zm5,0l0,1l-1,0l0,-1l1,0zm5,0l0,1l-1,0l0,-1l1,0zm5,0l0,1l-1,0l0,-1l1,0zm5,0l0,1l-1,0l0,-1l1,0zm5,0l0,1l-1,0l0,-1l1,0z"
+      />
+      <path
+        fillRule="nonzero"
+        fill="#8D91A3"
+        id="Path"
+        d="m266.45052,323.94524l0.84691,0.53175l-4.11509,6.55334l0.257,0.475l7.73895,0.1138l-0.01465,0.9999l-9.1778,-0.13444l-0.88879,-0.01302l0.47267,-0.7528l4.8808,-7.77353zm1.01523,4.49669l0.47745,0.87866l-0.87867,0.47745l-0.47744,-0.87867l0.87866,-0.47744zm4.32561,-2.47025l0.50792,0.86141l-0.8614,0.50792l-0.50792,-0.8614l0.8614,-0.50793zm4.24359,-2.61547l0.53178,0.84689l-0.84688,0.53178l-0.53179,-0.84689l0.84689,-0.53178zm4.17336,-2.73582l0.55522,0.8317l-0.83171,0.55522l-0.55521,-0.8317l0.8317,-0.55522zm4.09575,-2.85143l0.5782,0.8159l-0.8159,0.5782l-0.5782,-0.8159l0.8159,-0.5782zm4.01565,-2.9652l0.60072,0.79947l-0.79947,0.60071l-0.60071,-0.79946l0.79946,-0.60072zm3.92417,-3.06996l0.6227,0.78246l-0.78247,0.6227l-0.62269,-0.78247l0.78246,-0.62269zm3.84384,-3.18328l0.64422,0.76484l-0.76484,0.64422l-0.64422,-0.76484l0.76484,-0.64422zm3.7522,-3.28744l0.66523,0.74664l-0.74664,0.66522l-0.66523,-0.74664l0.74664,-0.66522zm3.66249,-3.39333l0.68573,0.72786l-0.72786,0.68573l-0.68573,-0.72786l0.72786,-0.68573zm3.57025,-3.49732l0.70571,0.7085l-0.7085,0.70571l-0.70571,-0.7085l0.7085,-0.70571zm3.47153,-3.59526l0.72514,0.68859l-0.68859,0.72515l-0.72515,-0.68859l0.6886,-0.72515zm3.34023,-3.65679l0.74386,0.66834l-0.66833,0.74386l-0.74386,-0.66833l0.66833,-0.74387zm3.26471,-3.77919l0.76217,0.64738l-0.64739,0.76217l-0.76216,-0.64739l0.64738,-0.76216zm3.15605,-3.86575l0.77987,0.62594l-0.62594,0.77987l-0.77987,-0.62594l0.62594,-0.77987zm3.04896,-3.95435l0.797,0.60398l-0.60398,0.797l-0.797,-0.60398l0.60398,-0.797zm2.93637,-4.03605l0.81352,0.58154l-0.58154,0.81352l-0.81352,-0.58154l0.58154,-0.81352zm2.82513,-4.12001l0.82944,0.5586l-0.5586,0.82944l-0.82944,-0.5586l0.5586,-0.82944zm2.70562,-4.19202l0.84471,0.53523l-0.53524,0.84471l-0.8447,-0.53524l0.53523,-0.8447zm2.59344,-4.27614l0.85937,0.51136l-0.51136,0.85937l-0.85937,-0.51136l0.51136,-0.85937z"
+      />
+      <path
+        fillRule="nonzero"
+        fill="#8D91A3"
+        id="Path"
+        d="m306.1674,117.87463l0.39841,0.9172l-0.9172,0.39841l-0.39841,-0.9172l0.9172,-0.39841zm-2.11756,-4.5516l0.4318,0.90197l-0.90197,0.4318l-0.4318,-0.90197l0.90197,-0.4318zm-2.28297,-4.47091l0.46456,0.88554l-0.88555,0.46456l-0.46455,-0.88555l0.88554,-0.46455zm-2.44722,-4.38809l0.49666,0.86794l-0.86794,0.49667l-0.49667,-0.86795l0.86795,-0.49666zm-2.60639,-4.2962l0.52806,0.84921l-0.84921,0.52806l-0.52805,-0.84921l0.8492,-0.52806zm-2.75383,-4.18694l0.55861,0.82943l-0.82942,0.55862l-0.55862,-0.82943l0.82943,-0.55862zm-2.91264,-4.0954l0.58847,0.80852l-0.80851,0.58848l-0.58848,-0.80852l0.80852,-0.58848zm-3.05968,-3.98641l0.61752,0.78655l-0.78656,0.61752l-0.61752,-0.78656l0.78656,-0.61751zm-3.20333,-3.87332l0.64571,0.76358l-0.76358,0.64571l-0.64572,-0.76358l0.76359,-0.64571zm-3.34716,-3.75989l0.67307,0.73957l-0.73957,0.67307l-0.67308,-0.73957l0.73958,-0.67307zm-3.44743,-3.60107l0.69928,0.71484l-0.71485,0.69928l-0.69928,-0.71484l0.71485,-0.69928zm-3.60519,-3.50291l0.72478,0.68897l-0.68897,0.72479l-0.72479,-0.68897l0.68898,-0.72479zm-3.72678,-3.36675l0.74932,0.66221l-0.6622,0.74932l-0.74932,-0.66221l0.6622,-0.74932zm-3.84821,-3.23052l0.77288,0.63454l-0.63454,0.77289l-0.77289,-0.63454l0.63455,-0.77289zm-3.96584,-3.0905l0.79547,0.60599l-0.606,0.79547l-0.79547,-0.606l0.606,-0.79546zm-4.07963,-2.9467l0.81704,0.57658l-0.57658,0.81704l-0.81704,-0.57658l0.57658,-0.81704zm-6.18206,-9.22492l0.00273,1l-8.28961,0.02279l4.32395,7.07261l-0.85318,0.52162l-4.78786,-7.83112l-0.46366,-0.75838l0.88888,-0.00243l9.17875,-0.02509zm2.00131,6.43149l0.83753,0.54639l-0.54639,0.83753l-0.83753,-0.54639l0.54639,-0.83753zm-4.28645,-2.64269l0.85697,0.51538l-0.51537,0.85696l-0.85697,-0.51537l0.51537,-0.85697zm-4.38484,-2.43511l0.87291,0.48788l-0.48788,0.87291l-0.87291,-0.48787l0.48788,-0.87292z"
+      />
+    </g>
+  );
+}
+
+function ArrowsFromCreated({ state }: { state: string }) {
+  if (state === 'Created') {
+    return (
+      <g>
+        <path
+          fillRule="nonzero"
+          fill="#2F5FA7"
+          id="Path"
+          d="m97.28988,13.52918l-5.69891,8.29291l-1.92923,-3.48854l-1.17306,0.6926c-53.58725,31.96892 -87.16394,89.75997 -87.16394,153.3944c0,66.81296 37.03273,127.03854 94.91512,157.76122l1.76038,0.92193l-0.45904,0.88842c-59.21997,-30.59866 -97.21646,-91.70607 -97.21646,-159.57157c0,-64.46443 34.26719,-122.9664 88.8524,-154.9633l-1.94094,-3.51145l10.05368,-0.41662z"
+        />
+        <path
+          fillRule="nonzero"
+          fill="#2F5FA7"
+          id="Path"
+          d="m297.36029,243.03748l-1.14161,9.99734l-3.36602,-2.19388l-0.27588,0.39536c-12.61953,17.91537 -29.36626,32.62019 -48.77879,42.78846l-0.89784,0.46608l-0.45745,-0.88924c19.78231,-10.17658 36.82327,-25.08068 49.5724,-43.30567l-3.33764,-2.17333l8.68283,-5.08512z"
+        />
+      </g>
+    );
+  }
+  return (
+    <g>
+      <path
+        d="M97.1134646,331.39154 L98.0002415,331.853738 L97.5380438,332.740515 L96.6512669,332.278317 L97.1134646,331.39154 Z M92.7120069,329.01826 L93.5854562,329.505175 L93.098541,330.378624 L92.2250918,329.891709 L92.7120069,329.01826 Z M88.3856482,326.526431 L89.2451219,327.037611 L88.7339418,327.897085 L87.8744682,327.385905 L88.3856482,326.526431 Z M84.1266146,323.911927 L84.971443,324.446965 L84.4364058,325.291793 L83.5915773,324.756756 L84.1266146,323.911927 Z M79.9474173,321.182693 L80.7769764,321.741112 L80.2185575,322.570671 L79.3889984,322.012252 L79.9474173,321.182693 Z M75.8406861,318.334086 L76.6543261,318.915455 L76.072957,319.729095 L75.259317,319.147726 L75.8406861,318.334086 Z M71.8185827,315.374617 L72.615701,315.978441 L72.0118777,316.775559 L71.2147595,316.171736 L71.8185827,315.374617 Z M67.8833173,312.306496 L68.6633225,312.932269 L68.0375495,313.712274 L67.2575443,313.086501 L67.8833173,312.306496 Z M64.028157,309.124337 L64.7904261,309.771597 L64.143166,310.533866 L63.3808968,309.886606 L64.028157,309.124337 Z M60.2734158,305.845294 L61.0174192,306.51347 L60.3492434,307.257473 L59.60524,306.589297 L60.2734158,305.845294 Z M56.6036461,302.456157 L57.3287776,303.144768 L56.6401672,303.869899 L55.9150357,303.181289 L56.6036461,302.456157 Z M53.0300169,298.966745 L53.7357217,299.675251 L53.0272158,300.380956 L52.3215109,299.67245 L53.0300169,298.966745 Z M49.554754,295.379284 L50.2404861,296.107138 L49.512632,296.792871 L48.8269,296.065016 L49.554754,295.379284 Z M46.1878855,291.704757 L46.8531556,292.451359 L46.1065529,293.116629 L45.4412828,292.370027 L46.1878855,291.704757 Z M42.9157883,287.928092 L43.5600192,288.692923 L42.7951882,289.337154 L42.1509573,288.572323 L42.9157883,287.928092 Z M39.7560383,284.069235 L40.3787594,284.851679 L39.5963156,285.4744 L38.9735944,284.691956 L39.7560383,284.069235 Z M36.7030425,280.121641 L37.3037425,280.921115 L36.5042681,281.521815 L35.903568,280.722341 L36.7030425,280.121641 Z M33.7590099,276.087519 L34.3371857,276.903431 L33.5212734,277.481607 L32.9430977,276.665695 L33.7590099,276.087519 Z M30.9522864,272.008228 L31.5076619,272.839828 L30.6760623,273.395203 L30.1206867,272.563603 L30.9522864,272.008228 Z M28.2567986,267.848347 L28.788897,268.695029 L27.9422145,269.227128 L27.4101161,268.380445 L28.2567986,267.848347 Z M25.6567048,263.579546 L26.1648876,264.440795 L25.3036384,264.948978 L24.7954556,264.087729 L25.6567048,263.579546 Z M23.1738067,259.233375 L23.6576096,260.108552 L22.7824326,260.592355 L22.2986297,259.717178 L23.1738067,259.233375 Z M20.8211611,254.833055 L21.280249,255.721446 L20.3918582,256.180534 L19.9327703,255.292143 L20.8211611,254.833055 Z M18.58894,250.360428 L19.0228759,251.261372 L18.1219321,251.695308 L17.6879962,250.794364 L18.58894,250.360428 Z M16.4890097,245.839347 L16.8974916,246.752114 L15.9847252,247.160595 L15.5762432,246.247829 L16.4890097,245.839347 Z M14.5126347,241.250913 L14.8952568,242.174818 L13.9713519,242.55744 L13.5887298,241.633535 L14.5126347,241.250913 Z M12.6662003,236.608411 L13.0226351,237.542731 L12.0883149,237.899166 L11.7318801,236.964846 L12.6662003,236.608411 Z M10.9548681,231.925715 L11.2848717,232.869695 L10.3408921,233.199698 L10.0108884,232.255719 L10.9548681,231.925715 Z M9.37525201,227.194343 L9.67853932,228.147243 L8.72564016,228.45053 L8.42235285,227.497631 L9.37525201,227.194343 Z M7.9351545,222.439946 L8.21159344,223.400978 L7.25056196,223.677417 L6.97412302,222.716385 L7.9351545,222.439946 Z M6.62173641,217.618926 L6.87095849,218.587373 L5.90251213,218.836595 L5.65329005,217.868148 L6.62173641,217.618926 Z M5.44606447,212.768433 L5.66792228,213.743512 L4.69284326,213.96537 L4.47098544,212.990291 L5.44606447,212.768433 Z M4.40822588,207.891225 L4.60260031,208.872152 L3.6216729,209.066527 L3.42729847,208.085599 L4.40822588,207.891225 Z M3.50621791,202.977925 L3.67295131,203.963927 L2.68694929,204.130661 L2.5202159,203.144659 L3.50621791,202.977925 Z M2.74615592,198.067377 L2.88532429,199.057646 L1.89505556,199.196814 L1.75588718,198.206545 L2.74615592,198.067377 Z M2.11910489,193.101376 L2.23047444,194.095155 L1.23669538,194.206524 L1.12532583,193.212745 L2.11910489,193.101376 Z M1.63318651,188.143437 L1.71689692,189.139928 L0.720406799,189.223638 L0.636696388,188.227148 L1.63318651,188.143437 Z M1.28413125,183.15918 L1.34014865,184.157609 L0.341718857,184.213627 L0.28570146,183.215197 L1.28413125,183.15918 Z M1.07348986,178.175859 L1.10195107,179.175454 L0.102356172,179.203915 L0.073894959,178.20432 L1.07348986,178.175859 Z M1,173.196076 L1.0010744,174.196076 L0.001074973,174.19715 L-2.84217094e-14,173.197151 L1,173.196076 Z M0.0636679723,168.168249 L1.06331895,168.194667 L1.03690085,169.194318 L0.037249871,169.1679 L0.0636679723,168.168249 Z M0.265803881,163.145883 L1.26434335,163.19991 L1.21031623,164.19845 L0.211776762,164.144423 L0.265803881,163.145883 Z M0.606561356,158.135971 L1.60321832,158.217671 L1.52151828,159.214328 L0.524861318,159.132628 L0.606561356,158.135971 Z M1.08604583,153.14092 L2.08004317,153.250324 L1.97063898,154.244321 L0.976641638,154.134917 L1.08604583,153.14092 Z M1.70594003,148.151477 L2.69648719,148.288649 L2.55931458,149.279196 L1.56876742,149.142024 L1.70594003,148.151477 Z M2.4633326,143.193702 L3.44965249,143.358544 L3.28480995,144.344864 L2.29849006,144.180022 L2.4633326,143.193702 Z M3.36174956,138.246853 L4.34304398,138.439366 L4.15053097,139.42066 L3.16923655,139.228147 L3.36174956,138.246853 Z M4.39691235,133.336685 L5.3724067,133.556709 L5.15238222,134.532204 L4.17688787,134.312179 L4.39691235,133.336685 Z M5.57063683,128.454279 L6.53954705,128.701692 L6.29213447,129.670602 L5.32322425,129.42319 L5.57063683,128.454279 Z M6.8795676,123.61353 L7.84113,123.888117 L7.56654348,124.849679 L6.60498108,124.575092 L6.8795676,123.61353 Z M8.32606191,118.805844 L9.27950178,119.107428 L8.97791862,120.060867 L8.02447875,119.759284 L8.32606191,118.805844 Z M9.90607858,114.044954 L10.8506467,114.373269 L10.5223311,115.317837 L9.577763,114.989522 L9.90607858,114.044954 Z M11.6225005,109.322615 L12.5574344,109.677436 L12.2026126,110.61237 L11.2676787,110.257549 L11.6225005,109.322615 Z M13.4662508,104.662965 L14.3908425,105.043925 L14.0098829,105.968516 L13.0852912,105.587557 L13.4662508,104.662965 Z M15.4445138,100.047241 L16.3580181,100.45407 L15.9511891,101.367574 L15.0376848,100.960745 L15.4445138,100.047241 Z M17.551931,95.4887724 L18.4536362,95.9211239 L18.0212846,96.8228291 L17.1195795,96.3904775 L17.551931,95.4887724 Z M19.7507594,91.0610306 L20.6401675,91.5181447 L20.1830535,92.4075527 L19.2936454,91.9504387 L19.7507594,91.0610306 Z M22.1106141,86.6238085 L22.9868406,87.105708 L22.5049411,87.9819345 L21.6287146,87.500035 L22.1106141,86.6238085 Z M24.5956358,82.2517696 L25.4579991,82.7580597 L24.9517089,83.6204229 L24.0893457,83.1141328 L24.5956358,82.2517696 Z M27.1984237,77.9572336 L28.0462849,78.4874518 L27.5160667,79.335313 L26.6682055,78.8050948 L27.1984237,77.9572336 Z M29.9230445,73.7329752 L30.7557433,74.2867013 L30.2020172,75.1194002 L29.3693184,74.5656741 L29.9230445,73.7329752 Z M32.7615418,69.5909888 L33.5784627,70.1677385 L33.001713,70.9846594 L32.1847921,70.4079096 L32.7615418,69.5909888 Z M35.7117029,65.5334856 L36.5122422,66.1327658 L35.912962,66.9333051 L35.1124227,66.3340249 L35.7117029,65.5334856 Z M38.778328,61.553833 L39.5618546,62.1751912 L38.9404965,62.9587178 L38.1569698,62.3373596 L38.778328,61.553833 Z M41.9526706,57.6634916 L42.7186005,58.3064158 L42.0756763,59.0723456 L41.3097465,58.4294215 L41.9526706,57.6634916 Z M45.2325038,53.8646874 L45.9802629,54.5286575 L45.3162928,55.2764166 L44.5685337,54.6124465 L45.2325038,53.8646874 Z M48.5924527,50.1843049 L49.3216049,50.8686565 L48.6372534,51.5978087 L47.9081012,50.9134571 L48.5924527,50.1843049 Z M52.0759137,46.5745977 L52.7857793,47.2789349 L52.0814422,47.9888005 L51.3715766,47.2844634 L52.0759137,46.5745977 Z M55.6663608,43.0553115 L56.356348,43.7791331 L55.6325264,44.4691203 L54.9425392,43.7452987 L55.6663608,43.0553115 Z M59.3454631,39.6444581 L60.0150776,40.3871669 L59.2723688,41.0567814 L58.6027543,40.3140726 L59.3454631,39.6444581 Z M63.1275192,36.3289449 L63.7761833,37.0900197 L63.0151085,37.7386838 L62.3664444,36.977609 L63.1275192,36.3289449 Z M66.9845504,33.1329492 L67.6118371,33.9117377 L66.8330486,34.5390243 L66.205762,33.7602359 L66.9845504,33.1329492 Z M70.9310958,30.0435833 L71.5364931,30.8395067 L70.7405696,31.444904 L70.1351724,30.6489805 L70.9310958,30.0435833 Z M74.9649444,27.0630592 L75.5479485,27.8755284 L74.7354793,28.4585325 L74.1524752,27.6460632 L74.9649444,27.0630592 Z M79.0838853,24.1935894 L79.6440007,25.022004 L78.8155861,25.5821194 L78.2554707,24.7537048 L79.0838853,24.1935894 Z M83.276152,21.4434656 L83.8129452,22.2871795 L82.9692313,22.8239726 L82.4324381,21.9802587 L83.276152,21.4434656 Z M87.7913008,13.5291757 L96.9697313,13.6092594 L97.8585864,13.6170148 L97.390387,14.3726025 L92.5557002,22.1748973 L91.7056641,21.648173 L95.780894,15.069 L95.520894,14.596 L87.782576,14.5291376 L87.7913008,13.5291757 Z M87.5487302,18.8082975 L88.0617257,19.6666888 L87.2033345,20.1796843 L86.6903389,19.3212931 L87.5487302,18.8082975 Z M91.9070407,16.3258405 L92.3896789,17.2016603 L91.5138591,17.6842985 L91.0312209,16.8084787 L91.9070407,16.3258405 Z"
+        id="Path"
+        fill="#8D91A3"
+      />
+      <path
+        d="M242.744167,294.000212 L243.201615,294.889448 L242.312379,295.346896 L241.854931,294.45766 L242.744167,294.000212 Z M247.11418,291.608517 L247.603849,292.480425 L246.73194,292.970094 L246.242271,292.098186 L247.11418,291.608517 Z M251.394111,289.058619 L251.915296,289.912063 L251.061852,290.433247 L250.540667,289.579803 L251.394111,289.058619 Z M255.584795,286.349844 L256.136799,287.183685 L255.302957,287.735689 L254.750954,286.901847 L255.584795,286.349844 Z M259.674799,283.489322 L260.256848,284.302476 L259.443694,284.884525 L258.861645,284.071371 L259.674799,283.489322 Z M263.643639,280.492818 L264.254815,281.284313 L263.46332,281.895489 L262.852144,281.103994 L263.643639,280.492818 Z M267.510457,277.346551 L268.149993,278.115313 L267.381231,278.754848 L266.741696,277.986087 L267.510457,277.346551 Z M271.267297,274.055708 L271.934375,274.800696 L271.189387,275.467774 L270.52231,274.722786 L271.267297,274.055708 Z M274.900656,270.631726 L275.59437,271.351976 L274.87412,272.045691 L274.180406,271.32544 L274.900656,270.631726 Z M278.409587,267.075446 L279.129034,267.769993 L278.434487,268.48944 L277.71504,267.794893 L278.409587,267.075446 Z M281.754548,263.43043 L282.498543,264.098615 L281.830358,264.84261 L281.086363,264.174425 L281.754548,263.43043 Z M284.993286,259.634117 L285.761075,260.274819 L285.120373,261.042608 L284.352583,260.401906 L284.993286,259.634117 Z M288.088681,255.726655 L288.879246,256.339035 L288.266866,257.129599 L287.476302,256.517219 L288.088681,255.726655 Z M291.044814,251.702315 L291.857168,252.285481 L291.274002,253.097834 L290.461649,252.514669 L291.044814,251.702315 Z M297.326861,243.037483 L297.274894,243.849 L297.360288,243.904598 L297.261894,244.056 L296.691531,253.085085 L295.693524,253.021979 L296.215894,244.748 L288.87459,248.597794 L288.410267,247.712128 L296.422894,243.511 L296.521388,243.360312 L296.605894,243.415 L297.326861,243.037483 Z M293.79996,247.554812 L294.63886,248.099097 L294.094574,248.937997 L293.255674,248.393711 L293.79996,247.554812 Z"
+        id="Path"
+        fill="#8D91A3"
+      />
+    </g>
+  );
+}
+
+function ArrowsFromDeprecated({ state }: { state: string }) {
+  if (state === 'Deprecated') {
+    return (
+      <g>
+        <polygon
+          points="629.519775390625,167.93579049408436 638.519775390625,172.43580575287342 629.519775390625,176.93579049408436 629.519775390625,172.93579049408436 596.519775390625,172.93579049408436 596.519775390625,171.93579049408436 629.519775390625,171.93579049408436"
+          fillRule="nonzero"
+          fill="#2F5FA7"
+          id="Path-3"
+        />
+      </g>
+    );
+  }
+  return (
+    <g>
+      <path
+        d="M629.745025,167.564212 L637.745025,172.064212 L638.519758,172.5 L637.745025,172.935788 L629.745025,177.435788 L629.254763,176.564212 L635.590894,173 L635.499894,173 L635.499894,172 L635.590894,172 L629.254763,168.435788 L629.745025,167.564212 Z M596.499894,172 L596.499894,173 L595.499894,173 L595.499894,172 L596.499894,172 Z M601.499894,172 L601.499894,173 L600.499894,173 L600.499894,172 L601.499894,172 Z M606.499894,172 L606.499894,173 L605.499894,173 L605.499894,172 L606.499894,172 Z M611.499894,172 L611.499894,173 L610.499894,173 L610.499894,172 L611.499894,172 Z M616.499894,172 L616.499894,173 L615.499894,173 L615.499894,172 L616.499894,172 Z M621.499894,172 L621.499894,173 L620.499894,173 L620.499894,172 L621.499894,172 Z M626.499894,172 L626.499894,173 L625.499894,173 L625.499894,172 L626.499894,172 Z M631.499894,172 L631.499894,173 L630.499894,173 L630.499894,172 L631.499894,172 Z"
+        id="Path-3"
+        fill="#8D91A3"
+      />
+    </g>
+  );
+}
+
+export interface LCStateDiagramProps {
+  currentState: string;
+  availableTransitions: LifecycleStateTransition[];
+}
+
+export default function LCStateDiagram({ currentState, availableTransitions }: LCStateDiagramProps): JSX.Element {
+  const state = currentState as ApiState;
+  return (
+    <svg width="780px" height="484px" viewBox="0 0 780 412" version="1.1" xmlns="http://www.w3.org/2000/svg" xmlnsXlink="http://www.w3.org/1999/xlink">
+      <defs>
+        <path
+          d="M56,3.46410162 L87.3012702,21.5358984 C91.0140831,23.6794919 93.3012702,27.6410162 93.3012702,31.9282032 L93.3012702,68.0717968 C93.3012702,72.3589838 91.0140831,76.3205081 87.3012702,78.4641016 L56,96.5358984 C52.2871871,98.6794919 47.7128129,98.6794919 44,96.5358984 L12.6987298,78.4641016 C8.98591689,76.3205081 6.69872981,72.3589838 6.69872981,68.0717968 L6.69872981,31.9282032 C6.69872981,27.6410162 8.98591689,23.6794919 12.6987298,21.5358984 L44,3.46410162 C47.7128129,1.32050808 52.2871871,1.32050808 56,3.46410162 Z"
+          id="path-1"
+        />
+        {SHADOW_FILTER}
+      </defs>
+      <g id="API" stroke="none" strokeWidth="1" fill="none" fillRule="evenodd">
+        <g id="Portal-Lifecycle" transform="translate(-399.000000, -160.000000)">
+          <g id="Group-10" transform="translate(399.500106, 160.000000)">
+            <g id="Group-9" transform="translate(0.000000, 31.000000)" fillRule="nonzero">
+              <polygon id="Path-3" fill="rgba(0,0,0,0.02)" points="709.440839 156.000116 709.558949 185.999884 328.558949 187.499884 328.440839 157.500116" />
+              <path
+                d="M179.499894,0 C151.196806,0 123.841338,6.81508263 99.29886,19.6770002 C42.7216813,49.3272661 6.49989404,107.956305 6.49989404,173 C6.49989404,268.545262 83.9546323,346 179.499894,346 C275.045156,346 352.499894,268.545262 352.499894,173 C352.499894,77.4547383 275.045156,0 179.499894,0 Z M179.499894,30 C258.476613,30 322.499894,94.0232808 322.499894,173 C322.499894,251.976719 258.476613,316 179.499894,316 C100.523175,316 36.499894,251.976719 36.499894,173 C36.499894,119.223698 66.4353632,70.7697715 113.224451,46.2491271 C133.495713,35.6256153 156.076707,30 179.499894,30 Z"
+                id="Oval"
+                fill="rgba(0,0,0,0.02)"
+              />
+              <ArrowsFromPrototyped state={state} />
+              <ArrowsFromPublished state={state} />
+              <ArrowsFromCreated state={state} />
+              <ArrowsFromDeprecated state={state} />
+            </g>
+            <PrototypedBubble style={getBubbleStyle('Prototyped', state, availableTransitions)} />
+            <PublishedBubble style={getBubbleStyle('Published', state, availableTransitions)} />
+            <CreatedBubble style={getBubbleStyle('Created', state, availableTransitions)} />
+            <DeprecatedBubble style={getBubbleStyle('Deprecated', state, availableTransitions)} />
+            <RetiredBubble style={getBubbleStyle('Retired', state, availableTransitions)} />
+          </g>
+        </g>
+      </g>
+    </svg>
+  );
+}
